pyproject.toml
[tool.poetry]
name = "sanders-customer-platform-cdk-infra"
version = "0.1.0"
description = "Sanders Customer Platform Infrastructure using AWS CDK"
authors = ["Sanders Team"]
classifiers = ["Private :: Do not Upload"]

[tool.poetry.dependencies]
python = ">=3.9,<=3.12"
aws-cdk-lib = "2.143.1"
constructs = ">=10.0.0,<11.0.0"

[tool.poetry.dev-dependencies]
pytest = "^7.4.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
